@@ -1,44 +1,49 @@
 {
-  "name": "ind21-poker-planning-nk-ug",
-  "version": "0.0.1",
-  "description": "Poker Planer",
+  "name": "planpoker",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest",
+    "watch": "ts-node-dev src/server.ts",
+    "migration:generate": "npx typeorm-ts-node-commonjs migration:generate ./src/migration/PostRefactoring -d ./src/settings/ormconfig.ts",
+    "migration:run": "npx typeorm-ts-node-commonjs migration:run -d ./src/settings/ormconfig.ts",
+    "migration:revert": "npx typeorm migration:revert -- -d ./src/settings/ormconfig.ts"
+  },
+  "author": "Utkarsh Goel",
+  "license": "ISC",
   "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
-    "@types/express": "^4.17.13",
-    "@types/hapi__joi": "^17.1.7",
-    "@types/json2csv": "^5.0.3",
-    "@types/jsonwebtoken": "^8.5.5",
-    "@types/node": "^8.0.29",
-    "@types/nodemailer": "^6.4.4",
-    "@types/uuid": "^8.3.1",
-    "install": "^0.13.0",
-    "npm": "^8.1.0",
-    "ts-node": "3.3.0",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5"
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.11",
+    "@types/jsonwebtoken": "^9.0.5",
+    "@types/node": "^20.10.6",
+    "@types/nodemailer": "^6.4.14",
+    "@typescript-eslint/eslint-plugin": "^6.17.0",
+    "@typescript-eslint/parser": "^6.17.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.2",
+    "eslint-plugin-tsdoc": "^0.2.17",
+    "jest": "^29.7.0",
+    "prettier": "^3.1.1",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@hapi/joi": "^17.1.1",
-    "@types/nodemailer": "^6.4.4",
-    "axios": "^0.21.3",
-    "bcrypt": "^5.0.1",
+    "argon2": "^0.31.2",
+    "axios": "^1.6.5",
     "cors": "^2.8.5",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "json2csv": "^5.0.6",
-    "jsonwebtoken": "^8.5.1",
-    "mysql": "^2.14.1",
-    "nodemailer": "^6.6.3",
-    "reflect-metadata": "^0.1.10",
-    "socket.io": "^4.1.3",
-    "socket.io-client": "^4.1.3",
-    "typeorm": "^0.2.34",
-    "uuid": "^8.3.2"
-  },
-  "scripts": {
-    "start": "ts-node-dev src/app.ts",
-    "migration:generate": "ts-node ./node_modules/typeorm/cli.js --config src/config/ormconfig.ts migration:generate -n",
-    "migration:run": "ts-node ./node_modules/typeorm/cli.js --config src/config/ormconfig.ts migration:run",
-    "migration:revert": "ts-node ./node_modules/typeorm/cli.js --config src/config/ormconfig.ts migration:revert"
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "http-status-codes": "^2.3.0",
+    "jsonwebtoken": "^9.0.2",
+    "mysql": "^2.18.1",
+    "nodemailer": "^6.9.8",
+    "reflect-metadata": "^0.2.1",
+    "typeorm": "^0.3.17",
+    "zod": "^3.22.4"
   }
 }
