@@ -1,155 +1,33 @@
 import { Request, Response } from 'express';
-import { getRepository } from 'typeorm';
-import { ErrorTypes } from '../constants/errorType';
-import { ConstantValues } from '../constants/constantValues';
-import { ErrorMessage, SuccessMessage, TEXT } from '../constants/message';
-import { SEARCH_LIMIT, SortType, SORT_TYPE } from '../constants/customTypes';
-import { Pokerboard } from '../entity/Pokerboard';
-import { Ticket } from '../entity/Ticket';
-import { UserPokerboard } from '../entity/UserPokerboard';
-import { UserTicket } from '../entity/UserTicket';
-import { generateCustomResponse } from '../middlewares/user.validation';
-import { IRequest } from '../models/RequestInterface';
-import ResponseMessage from '../models/ResponseMessage';
-import {
-  getGroupDetailsAssociatedWithUser,
-  userGroupData,
-} from '../repositories/group.repository';
-import { GroupDetails } from '../repositories/pokerBoard.repository';
-import {
-  findUser,
-  getPokerboardDetailsAssociatedWithUser,
-  getReportOfUserHelper,
-  PokerboardDetailsType,
-} from '../repositories/user.repository';
+import { StatusCodes } from 'http-status-codes';
+import { FindManyOptions, Like, Not } from 'typeorm';
 
-export const searchUser = async (req: Request, res: Response) => {
-  let { searchKey, limit: limitFromRequest } = req.query;
-  if (!searchKey) {
-    searchKey = '';
-  }
-  const limit = Number(limitFromRequest)
-    ? Number(limitFromRequest)
-    : SEARCH_LIMIT;
+import { ONE_DAY, UNDEFINED } from '../constants/common';
+import { SortTypes } from '../constants/enums';
+import { ErrorMessages, ResponseMessages, ValidationMessages } from '../constants/message';
+import { Ticket } from '../entity/ticket/model';
+import { User } from '../entity/user/model';
+import { findUser, findUsers, saveUser } from '../entity/user/repository';
+import { UserTicket } from '../entity/userTicket/model';
+import { makeResponse } from '../utils/common';
+import customGetRepository from '../utils/db';
 
-  let users = await findUser(searchKey.toString(), limit);
-
-  let searchResult = users.map((user) => {
-    let {
-      email,
-      userName: user_name,
-      firstName: first_name,
-      lastName: last_name,
-      id,
-    } = user;
-    return {
-      email,
-      userName: user_name,
-      name: `${first_name} ${last_name}`,
-      id,
-    };
-  });
-
-  const message: ResponseMessage = {
-    data: searchResult,
-    success: true,
-    message: '',
-  };
-  return res.json(message);
-};
-
-export const getPokerboardAssociatedToUser = async (
-  req: IRequest,
-  res: Response
-) => {
+export const getTicketsAssociatedToUser = async (req: Request, res: Response) => {
   const user = req.user;
+  let { filter, date } = req.query;
 
-  const userPokerboards: PokerboardDetailsType[] =
-    await getPokerboardDetailsAssociatedWithUser(user);
+  const newSort =
+    req.query.sort == SortTypes.ASCENDING ? SortTypes.ASCENDING : SortTypes.DESCENDING;
+  filter = filter == UNDEFINED ? undefined : filter?.toString();
 
-  res
-    .status(200)
-    .json(
-      generateCustomResponse(
-        true,
-        SuccessMessage.FETCHED_DETAILS_SUCCESSFULLY,
-        userPokerboards
-      )
-    );
-};
-
-export const getGroupsAssociatedToUser = async (
-  req: IRequest,
-  res: Response
-) => {
-  let responseMessage: ResponseMessage = {
-    data: undefined,
-    message: '',
-    success: true,
-  };
-
-  const user = req.user;
-  let groups = await user.groups;
-  let groupIds = groups.map((group) => group.id);
-
-  let groupsData: GroupDetails[] = await getGroupDetailsAssociatedWithUser(
-    groupIds
-  );
-
-  let userGroups: { [key: string]: userGroupData } = {};
-  for (let groupData of groupsData) {
-    if (!userGroups[groupData.id]) {
-      userGroups[groupData.id] = {
-        id: groupData.id,
-        name: groupData.name,
-        admin: groupData.admin,
-        countOfMembers: groupData.countOfMembers,
-        members: [],
-      };
-    }
-    userGroups[groupData.id].members.push(groupData.member);
-  }
-
-  responseMessage.data = {
-    groups: Object.values(userGroups),
-  };
-  res.status(200).json(responseMessage);
-};
-
-export type ticketDetails = {
-  id: string;
-  type: string;
-  managerEstimate: number;
-  playerEstimate: number;
-  date: Date;
-};
-
-export const getTicketsAssociatedToUser = async (
-  req: IRequest,
-  res: Response
-) => {
-  let responseMessage: ResponseMessage = {
-    data: undefined,
-    message: '',
-    success: true,
-  };
-
-  const user = req.user;
-  let { sort, filter, date } = req.query;
-
-  const newSort: SORT_TYPE =
-    sort == SortType.ASCENDING ? SortType.ASCENDING : SortType.DESCENDING;
-  filter = filter == TEXT.undefined ? undefined : filter.toString();
-  date = date == TEXT.undefined ? undefined : date.toString();
-
+  date = date == UNDEFINED ? undefined : date?.toString();
   let from: Date, to: Date;
-
   if (date) {
     from = new Date(+date);
-    to = new Date(+date + ConstantValues.oneDay);
+    to = new Date(+date + ONE_DAY);
   }
 
-  let tickets = await getRepository(UserTicket)
+  const tickets = await customGetRepository(UserTicket)
     .createQueryBuilder('user_ticket')
     .where({
       userId: user.id,
@@ -163,58 +41,88 @@ export const getTicketsAssociatedToUser = async (
     .orderBy('date', newSort)
     .andWhere(filter ? `type = :filter` : '1=1', { filter })
     .andWhere(
-      date
-        ? `user_ticket.estimate_date >= :from AND user_ticket.estimate_date <= :to`
-        : '1=1',
-      {
-        from: from,
-        to: to,
-      }
+      date ? `user_ticket.estimate_date >= :from AND user_ticket.estimate_date <= :to` : '1=1',
+      { from: from, to: to },
     )
     .getRawMany();
 
-  responseMessage.data = {
-    ticketDetails: tickets,
-  };
-
-  res.status(200).json(responseMessage);
+  res
+    .status(StatusCodes.OK)
+    .json(makeResponse(true, ResponseMessages.GET_TICKETS_SUCCESS, { ticketDetails: tickets }));
 };
 
-export const getReportOfUser = async (req: IRequest, res: Response) => {
-  let pokerboard: Pokerboard = req.pokerboard;
-  let userRole: UserPokerboard = req.userRole;
-  let user = req.user;
-
-  try {
-    let report = await getReportOfUserHelper(pokerboard, user, userRole);
+export const getUser = async (req: Request, res: Response) => {
+  const { password, ...user } = req.user; // eslint-disable-line @typescript-eslint/no-unused-vars
+  if (!user.isVerified) {
+    const errorData = { verify: ErrorMessages.ACCOUNT_NOT_VERIFIED };
     return res
-      .status(200)
-      .json(
-        generateCustomResponse(
-          true,
-          SuccessMessage.FETCHED_DETAILS_SUCCESSFULLY,
-          report
-        )
-      );
-  } catch (error) {
-    let errData: any = {};
-    if (error instanceof Error) {
-      errData[error.name] = error.message;
+      .status(StatusCodes.BAD_REQUEST)
+      .json(makeResponse(false, ErrorMessages.ACCOUNT_NOT_VERIFIED, errorData));
+  } else {
+    return res
+      .status(StatusCodes.OK)
+      .json(makeResponse(true, ResponseMessages.GET_USER_SUCCESS, user));
+  }
+};
+
+export const searchUsers = async (req: Request, res: Response) => {
+  const searchKey = req.query.searchKey.toString();
+  const limit = Number(req.query.limit);
+  const findCondition = { isActive: true, isVerified: true };
+  const findOptions: FindManyOptions<User> = {
+    where: [
+      { ...findCondition, username: Like(`${searchKey}%`) },
+      { ...findCondition, email: Like(`${searchKey}%`) },
+      { ...findCondition, firstName: Like(`${searchKey}%`) },
+      { ...findCondition, lastName: Like(`${searchKey}%`) },
+    ],
+    select: ['email', 'firstName', 'lastName', 'id', 'username'],
+    take: limit,
+  };
+
+  const users = await findUsers(findOptions);
+
+  const searchResult = users.map((user) => {
+    return {
+      email: user.email,
+      username: user.username,
+      name: `${user.firstName} ${user.lastName}`,
+      id: user.id,
+    };
+  });
+
+  res
+    .status(StatusCodes.OK)
+    .json(makeResponse(true, ResponseMessages.USER_SEARCH_SUCCESS, searchResult));
+};
+
+export const updateUser = async (req: Request, res: Response) => {
+  const { firstName, lastName, username } = req.body;
+  const user = req.user;
+
+  if (firstName) {
+    user.firstName = firstName;
+  }
+
+  if (lastName) {
+    user.lastName = lastName;
+  }
+
+  if (username) {
+    const userWithSameUsername = await findUser({ username: username, id: Not(user.id) });
+    if (userWithSameUsername) {
+      const errorData = { username: ValidationMessages.USERNAME_ALREADY_EXISTS };
       return res
-        .status(400)
-        .json(generateCustomResponse(false, error.message, errData));
+        .status(StatusCodes.BAD_REQUEST)
+        .json(makeResponse(false, ValidationMessages.USERNAME_ALREADY_EXISTS, errorData));
     } else {
-      errData[ErrorTypes.SOME_THING_WENT_WRONG] =
-        ErrorMessage.SOMETHING_WENT_WRONG;
-      return res
-        .status(500)
-        .json(
-          generateCustomResponse(
-            false,
-            ErrorMessage.SOMETHING_WENT_WRONG,
-            errData
-          )
-        );
+      user.username = username;
     }
   }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const { password, ...userData } = await saveUser(user);
+  return res
+    .status(StatusCodes.OK)
+    .json(makeResponse(true, ResponseMessages.UPDATE_USER_SUCCESS, userData));
 };
